Add sl and faith to notnics
sl0 and faith0 use to be present in DragonFly BSD and they are not supposed to work like NICs so add them to notnics. I think this won't hurt any other BSD.

Also stuff like tap or bridge would need to be in notnics but for different reasons. In my opinion, what would be really needed is a configuration dialog.